requirements.txt
uvicorn[standard]
fastapi
pydantic
pydantic_settings
ipykernel
sqlalchemy
pyodbc
google-genai
mcp
alembic